Refactoring partitioned-heat-conduction/nutils (#295)
* forward compatibility

This patch removes the leading dashes for cli arguments in the run script,
which were always optional but have recently been forbidden. The changes are
compatible with all versions of nutils.

* disable rich output

This patch disables rich output by default, in order for standard output
logging to co-exist nicely with precice's log messages.

* remove droptol for projection of initial condition

This patch removes the droptol argument in the projection of the initial
condition, which is a full domain projection that should include all degrees of
freedom.

* simplify definition of sqr0

This patch simplifies the boundary constraints by using uexact.

* switch from asfunction to sample.basis

This patch replaces sample.asfunction by sample.basis combined with an argument
for readdata. While both constructions yield the same object internally
(asfunction uses basis under the hood), the new form makes it possible to
construct functionals ahead of time and benefit from cached data structures,
rather that recreate them at every iteration (see next commit).

* remove redefinition of sqr and res in time loop

This patch defines res and sqr outside the time loop, so that the only things
changing per timestep are the arguments (including readdata). The resulting
code is both cleaner and faster due to reuse of cached structures.

* introduce precice_read, precice_write

This patch simplifies the read and write instructions by sticking the arguments
inside functools.partial, leaving a more readable remainder of the code.

* improve flux evaluation

This patch replaces the projection matrix by a functional based equivalent that
is a bit more ideomatic nutils. It also fixes a consistency issue at the ends
of the boundary and adds an elaborate comment to explain the construction.

* streamline checkpointing

This patch bundles the checkpointed state in a single `checkpoint' tuple and
unifies output if the initial condition and following time steps.

* make parameters into function arguments

This patch makes all parameters into arguments of main so that they can be
configured from the command line (though not via run.sh). This makes it
possible, for example, to experiment with different timesteps on either half of
the simulation.