Rename participant and meshes
Solid1 becomes Solid-Left, and Solid2 becomes Solid-Right following the directory names.
We could further rename these, together with the directory names.

For the Fluid meshes, we just use Upstream and Downstream as indicators.

For the Solid meshes, we follow the usual Participant-Mesh name.

The data fields are named based on the Upstream/Downstream.