Add Nutils participants for turek-hron-fsi3 (#648)
* Add Nutils participants for turek-hron-fsi3

* Add python requirements

* Change Nutils cases from FSI2 to FSI3

* Remove dummy mock from solid.py

* Adjust README and plotting script

* Update turek-hron-fsi3/README.md

Co-authored-by: Gerasimos Chourdakis <gerasimos.chourdakis@ipvs.uni-stuttgart.de>

* Restructure time loop

This patch conforms the code to the standard control flow order of
requires_writing_checkpoint, requires_reading_checkpoint and
is_time_window_complete.

* Prepare for dynamic timesteps

This patch removes the Dynamic.set_timestep method, which is not a good fit in
a situation where time steps may vary. Instead the add_and_plot method is
rewritten to use dimensional times rather than a fixed-length deque, and the
newmark_defo and newmark_velo methods receive a timestep argument to replace
the attribute. The latter will be further modified in a subsequent commit to
allow for truly dynamic time steps.

* Rename t->traction to make place for time variable

* Update precice timestep at every iteration

* Acknowledgements

---------

Co-authored-by: Benjamin Uekermann <benjamin.uekermann@ipvs.uni-stuttgart.de>
Co-authored-by: Benjamin Uekermann <benjamin.uekermann@gmail.com>
Co-authored-by: Gerasimos Chourdakis <gerasimos.chourdakis@ipvs.uni-stuttgart.de>